Making integration adjustments

+++ lib/globals.dart
@@ -2,6 +2,6 @@ class Global {
   static int playerId;
   static int currentGameId;
   static int roundId;
-  static int roundBet;
+  static double roundBet;
   static int playerTurnId;
 }
